# docs/swagger.yaml
basePath: /
definitions:
...
  web.TargetPokinOpdTacticalCreateRequest:
    properties:
      indikator_pokin_opd_tactical_id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    required:
    - indikator_pokin_opd_tactical_id
    - satuan
    - target
    type: object
  web.TargetPokinOpdTacticalResponse:
    properties:
      indikator_pokin_opd_tactical_id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    type: object
  web.TargetPokinOpdTacticalUpdateRequest:
    properties:
      id:
        type: integer
      indikator_pokin_opd_tactical_id:
        type: integer
      satuan:
        type: string
      target:
        type: integer
    required:
    - id
    - indikator_pokin_opd_tactical_id
    - satuan
    - target
    type: object
...
  web.TujuanPokinOpdCreateRequest:
    properties:
      bidang_urusan:
        type: string
      kode_opd:
        type: string
      pokin_opd_id:
        type: integer
      tahun_akhir_periode:
        type: integer
      tahun_awal_periode:
        type: integer
      tujuan:
        type: string
    type: object
  web.TujuanPokinOpdResponse:
    properties:
      bidang_urusan:
        type: string
      id:
        type: integer
      kode_opd:
        type: string
      tahun_akhir_periode:
        type: integer
      tahun_awal_periode:
        type: integer
      tujuan:
        type: string
    type: object
  web.TujuanPokinOpdUpdateRequest:
    properties:
      bidang_urusan:
        type: string
      id:
        type: integer
      kode_opd:
        type: string
      tahun_akhir_periode:
        type: integer
      tahun_awal_periode:
        type: integer
      tujuan:
        type: string
    required:
    - id
    type: object
  web.WebResponse:
    description: Web Response
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: ${PROD_HOSTNAME}
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API For Alur Kerja Services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Alur Kerja Service API
  version: "1.0"
paths:
...
  /target-pokin-opd-tacticals:
    get:
      consumes:
      - application/json
      description: Get list of all Target Pokin OPD Tactical
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.TargetPokinOpdTacticalResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: List All Target Pokin OPD Tactical
      tags:
      - Target Pokin OPD Tactical
    post:
      consumes:
      - application/json
      description: Create new Target Pokin OPD Tactical
      parameters:
      - description: Target Pokin OPD Tactical Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TargetPokinOpdTacticalCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TargetPokinOpdTacticalResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Target Pokin OPD Tactical
      tags:
      - Target Pokin OPD Tactical
  /target-pokin-opd-tacticals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing Target Pokin OPD Tactical by ID
      parameters:
      - description: Target Pokin OPD Tactical ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Target Pokin OPD Tactical
      tags:
      - Target Pokin OPD Tactical
    get:
      consumes:
      - application/json
      description: Get Target Pokin OPD Tactical detail by ID
      parameters:
      - description: Target Pokin OPD Tactical ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TargetPokinOpdTacticalResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Target Pokin OPD Tactical by ID
      tags:
      - Target Pokin OPD Tactical
    put:
      consumes:
      - application/json
      description: Update existing Target Pokin OPD Tactical by ID
      parameters:
      - description: Target Pokin OPD Tactical ID
        in: path
        name: id
        required: true
        type: integer
      - description: Target Pokin OPD Tactical Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TargetPokinOpdTacticalUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TargetPokinOpdTacticalResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Target Pokin OPD Tactical
      tags:
      - Target Pokin OPD Tactical
...
  /tujuan-pokin-opds:
    get:
      consumes:
      - application/json
      description: Get list of all Tujuan Pokin Opd
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.TujuanPokinOpdResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: List All Tujuan Pokin Opd
      tags:
      - Tujuan Pokin Opd
    post:
      consumes:
      - application/json
      description: Create new Tujuan Pokin Opd
      parameters:
      - description: Tujuan Pokin Opd Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TujuanPokinOpdCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TujuanPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Tujuan Pokin Opd
      tags:
      - Tujuan Pokin Opd
  /tujuan-pokin-opds/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing Tujuan Pokin Opd by ID
      parameters:
      - description: Tujuan Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TujuanPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Tujuan Pokin Opd
      tags:
      - Tujuan Pokin Opd
    get:
      consumes:
      - application/json
      description: Get Tujuan Pokin Opd detail by ID
      parameters:
      - description: Tujuan Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TujuanPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Tujuan Pokin Opd by ID
      tags:
      - Tujuan Pokin Opd
    put:
      consumes:
      - application/json
      description: Update existing Tujuan Pokin Opd by ID
      parameters:
      - description: Tujuan Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tujuan Pokin Opd Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.TujuanPokinOpdUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.TujuanPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Tujuan Update Pokin Opd
      tags:
      - Tujuan Pokin Opd
schemes:
- http
- https
swagger: "2.0"